+++ tests/oauth2_connector/test_oauth2connector.py
@@ -11,6 +11,7 @@ from toucan_connectors.oauth2_connector.oauth2connector import (
     NoOAuth2RefreshToken,
     OAuth2Connector,
     OAuth2ConnectorConfig,
+    SecretKeeperMissingError,
 )
 from toucan_connectors.snowflake_oauth2.snowflake_oauth2_connector import SnowflakeoAuth2Connector
 from toucan_connectors.toucan_connector import get_oauth2_configuration
@@ -30,6 +31,18 @@ def oauth2_connector(secrets_keeper):
         redirect_uri="",
         token_url=FAKE_TOKEN_URL,
         secrets_keeper=secrets_keeper,
+    )
+
+
+@pytest.fixture
+def oauth2_connector_without_secret_keeper():
+    return OAuth2Connector(
+        auth_flow_id="test",
+        authorization_url=FAKE_AUTHORIZATION_URL,
+        scope=SCOPE,
+        config=OAuth2ConnectorConfig(client_id="", client_secret=""),
+        redirect_uri="",
+        token_url=FAKE_TOKEN_URL,
     )
 
 
@@ -234,3 +247,20 @@ def test_get_refresh_token(mocker, oauth2_connector):
     mocked_load.return_value = {"refresh_token": "bla"}
     token = oauth2_connector.get_refresh_token()
     assert token == "bla"
+
+
+def test_raises_exception_if_secret_keeper_not_set(oauth2_connector_without_secret_keeper: OAuth2Connector):
+    with pytest.raises(SecretKeeperMissingError):
+        oauth2_connector_without_secret_keeper.get_access_token()
+
+    with pytest.raises(SecretKeeperMissingError):
+        oauth2_connector_without_secret_keeper.retrieve_tokens(authorization_response="")
+
+    with pytest.raises(SecretKeeperMissingError):
+        oauth2_connector_without_secret_keeper.build_authorization_url()
+
+    with pytest.raises(SecretKeeperMissingError):
+        oauth2_connector_without_secret_keeper.get_access_data()
+
+    with pytest.raises(SecretKeeperMissingError):
+        oauth2_connector_without_secret_keeper.get_refresh_token()